Eliminated duplicated test code
